Initial naive implementation of timeseries parsing
A timeseries result can be bigint, double, boolean, or varchar. This commit just attempts to parse a boolean and a number and falls back to varchar as the default. 

Its not clear how timeseries views distribute across query tokens so this implementation will behave in unexpected ways if the query result is not the full timeseries view.  In my testing this has not been a problem returning views if ~1k time/value objects.